null check for book picture deleting was added.

+++ library-web/src/main/java/com/nix/libraryweb/model/repository/handler/BookEventHandler.java
@@ -19,6 +19,8 @@ public class BookEventHandler {
 
     @HandleAfterDelete
     public void deleteBookFile(Book book) {
-        fileService.delete(book.getPicturePath());
+        if (book.getPicturePath() != null) {
+            fileService.delete(book.getPicturePath());
+        }
     }
 }
